Improved install location define.
There's a single UGDK_INSTALL_LOCATION define that indicates where the
UGDK is installed in the system. Inside this directory should be one
different directory for each UGDK major version that is installed.
The enviroment variable UGDK_DIR could be set to the same directory as
the UGDK_INSTALL_LOCATION macro.

Python scripts are installed to UGDK_INSTALL_LOCATION/BIG_VERSION/python