AnimationManager massive changes.
The basic AnimationManager simply manages an abstract object of type "Animation".
It also only manages a single Animation, but handles changes to it's Animation.
SpecializedAnimationManager is a template extension to the AnimationManager that uses a IndexedTable for ease changing of Animations. The template argument must be a class that is a child of Animation.

The previous class Animation is now called SpriteAnimation and is a child of the new Animation. The AnimationFrame became a SpriteAnimationFrame.

The Sprite interface no longer wraps the AnimationManager: it has a getter for the animationmanager and that's enough.