Changing how creating UGDK executables work.
There's a new macro in the main UGDK CMakeLists called add_ugdk_executable that creates that given the target name and the sources, creates the executable and links it to UGDK.
However, in the case of an android build, that macro creates a shared lib and adds the extra source file that provides the necessary JNI hook.
Also, the examples have more meaningful names and should all be shared between desktop and android.